refactor: move link button
