+++ packages/components/buttons/link-button/src/link-button.story.js
@@ -2,9 +2,9 @@ import React from 'react';
 import { storiesOf } from '@storybook/react';
 import { withKnobs, boolean, text, select } from '@storybook/addon-knobs/react';
 import { BrowserRouter as Router } from 'react-router-dom';
-import Section from '../../../../.storybook/decorators/section';
-import * as icons from '../../icons';
-import Readme from './README.md';
+import * as icons from '@commercetools-uikit/icons';
+import Section from '../../../../../.storybook/decorators/section';
+import Readme from '../README.md';
 import LinkButton from './link-button';
 
 const iconNames = Object.keys(icons);
