+++ packages/components/buttons/link-button/src/link-button.visualroute.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { LinkButton, InformationIcon } from '@commercetools-frontend/ui-kit';
-import { Suite, Spec } from '../../../../test/percy';
+import { Suite, Spec } from '../../../../../test/percy';
 
 export const routePath = '/link-button';
